処理のカスタムフック化、コンポーネントに切り出し
