+++ src/App.js
@@ -1,110 +1,20 @@
-import { useState, useEffect } from "react";
+import useMemoManager from "./hooks/useMemoManager";
+import Detail from "./components/Detail";
+import List from "./components/List";
 import "./App.css";
 
-function List({ memos, selectedId, onSelected, onAdd }) {
-  const extractTitle = (memo) => memo.content.split("\n")[0];
-  const applySelectedClass = (id) => (id === selectedId ? "selected" : "");
-  return (
-    <div className="list">
-      <ul>
-        {memos.map((memo) => (
-          <li
-            key={memo.id}
-            className={applySelectedClass(memo.id)}
-            onClick={(e) => onSelected(memo.id, e)}
-          >
-            {extractTitle(memo)}
-          </li>
-        ))}
-        <li
-          onClick={(e) => {
-            e.stopPropagation();
-            onAdd();
-          }}
-          className={selectedId ? "disabled" : ""}
-        >
-          +
-        </li>
-      </ul>
-    </div>
-  );
-}
-
-function Detail({
-  inputContent,
-  onInputChange,
-  onUpdate,
-  onDelete,
-  selectedId,
-}) {
-  return (
-    <div className="detail">
-      <input
-        value={inputContent}
-        onChange={onInputChange}
-        onClick={(e) => e.stopPropagation()}
-        disabled={!selectedId}
-      />
-      <div className="action">
-        <button className="update" onClick={onUpdate} disabled={!selectedId}>
-          更新
-        </button>
-        <button className="delete" onClick={onDelete} disabled={!selectedId}>
-          削除
-        </button>
-      </div>
-    </div>
-  );
-}
-
 function App() {
-  const [selectedId, setSelectedId] = useState(null);
-  const [memos, setMemos] = useState(() => {
-    const storedMemos = JSON.parse(localStorage.getItem("memos"));
-    return storedMemos ? storedMemos : [];
-  });
-  const [inputContent, setInputContent] = useState("");
-  const selectMemo = (id, content) => {
-    setSelectedId(id);
-    setInputContent(content);
-  };
-
-  const handleAdd = () => {
-    if (selectedId) return;
-    const newMemo = { id: crypto.randomUUID(), content: "新規メモ" };
-    const newMemos = [...memos, newMemo];
-    selectMemo(newMemo.id, newMemo.content);
-    setMemos(newMemos);
-  };
-  const handleUpdate = () => {
-    if (!selectedId) return;
-    const updateMemo = {
-      id: memos.find((memo) => memo.id === selectedId).id,
-      content: inputContent,
-    };
-    const newMemos = memos.map((memo) =>
-      memo.id === selectedId ? updateMemo : memo,
-    );
-    selectMemo(null, "");
-    setMemos(newMemos);
-  };
-  const handleDelete = () => {
-    if (!selectedId) return;
-    const newMemos = memos.filter((memo) => memo.id !== selectedId);
-    selectMemo(null, "");
-    setMemos(newMemos);
-  };
-  const handleSelect = (id, e) => {
-    e.stopPropagation();
-    const selectedContent = memos.find((memo) => memo.id === id).content;
-    selectMemo(id, selectedContent);
-  };
-  const handleDeselect = () => selectMemo(null, "");
-  const handleInputChange = (e) => setInputContent(e.target.value);
-
-  useEffect(() => {
-    localStorage.setItem("memos", JSON.stringify(memos));
-  }, [memos]);
+  const {
+    selectedId,
+    memos,
+    inputContent,
+    handleAdd,
+    handleUpdate,
+    handleDelete,
+    handleSelect,
+    handleDeselect,
+    handleInputChange,
+  } = useMemoManager();
 
   return (
     <div className="app" onClick={handleDeselect}>
