+++ src/index.css
@@ -14,67 +14,3 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
     monospace;
 }
-
-.app {
-  width: 500px;
-}
-
-.index {
-  .list {
-    border: 1px solid;
-  }
-}
-
-.edit {
-  .list_detail {
-    border: 1px solid;
-    display: flex;
-    justify-content: space-between;
-  }
-}
-
-.list {
-  ul {
-    list-style: none;
-    li {
-      cursor: pointer;
-      margin: 20px 0;
-      width: fit-content;
-      &:not(.selected) {
-        color: blue;
-        text-decoration: underline;
-      }
-      &.selected {
-        color: inherit;
-        text-decoration: none;
-      }
-      &.disabled {
-        color: inherit;
-        text-decoration: none;
-        cursor: not-allowed;
-      }
-    }
-  }
-}
-
-.detail {
-  display: flex;
-  flex-direction: column;
-  padding: 20px;
-  width: 50%;
-  input {
-    height: 100px;
-  }
-  .action {
-    display: flex;
-    justify-content: space-between;
-    margin-top: 20px;
-    button {
-      width: 50%;
-      cursor: pointer;
-      &:disabled {
-        cursor: not-allowed;
-      }
-    }
-  }
-}
